Move repository module out of project
