+++ packages/ploys-api/src/github/webhook/mod.rs
@@ -7,8 +7,8 @@ pub mod secret;
 use axum::extract::State;
 use ploys::changelog::Changelog;
 use ploys::package::BumpOrVersion;
-use ploys::project::repository::revision::Revision;
 use ploys::project::Project;
+use ploys::repository::revision::Revision;
 use reqwest::Client;
 use semver::Version;
 use serde::{Deserialize, Serialize};
