+++ packages/ploys/src/lib.rs
@@ -2,3 +2,4 @@ pub mod changelog;
 pub mod lockfile;
 pub mod package;
 pub mod project;
+pub mod repository;
